+++ src/Resources/app/administration/src/main.js
@@ -2,5 +2,7 @@
 import './init/topdata-mapper-services.init';
 
 // ---- modules ----
+import './module/topdata-mapper';
+import './module/topdata-mapper-mappings';
 import './module/topdata-mapper-settings';
 import './module/topdata-mapper-conflicts';